+++ apps/docs/src/app/core/core-documentation.routes.ts
@@ -469,6 +469,10 @@ export const ROUTES: Routes = [
                     )
             },
             {
+                path: 'upload-collection',
+                loadChildren: () => import('./component-docs/upload-collection/upload-collection-docs.module').then((m) => m.UploadCollectionDocsModule)
+            },
+            {
                 path: 'wizard',
                 loadChildren: () => import('./component-docs/wizard/wizard-docs.module').then((m) => m.WizardDocsModule)
             },