+++ apps/docs/src/app/platform/component-docs/platform-wizard-generator/examples/wizard-generator-customizable-example.component.ts
@@ -16,6 +16,7 @@ export class WizardGeneratorCustomizableExampleComponent implements OnDestroy {
     @ViewChild('finishTemplate') finishTemplate: TemplateRef<any>;
     @ViewChild('cancelTemplate') cancelTemplate: TemplateRef<any>;
     @ViewChild('confirmationDialogTemplate') confirmationDialogTemplate: TemplateRef<any>;
+    @ViewChild('summaryStepTemplate') summaryStepTemplate: TemplateRef<any>;
 
     wizardValue: WizardGeneratorFormsValue;
 
@@ -146,11 +147,17 @@ export class WizardGeneratorCustomizableExampleComponent implements OnDestroy {
                 goBackButtonTemplate: this.goBackTemplate,
                 finishButtonTemplate: this.finishTemplate,
                 cancelButtonTemplate: this.cancelTemplate,
-                confirmationDialogTemplate: this.confirmationDialogTemplate
+                confirmationDialogTemplate: this.confirmationDialogTemplate,
+                summaryStepTemplate: this.summaryStepTemplate
             }
         }).afterClosed.pipe(takeUntil(this._onDestroy$))
         .subscribe((wizardValue: WizardGeneratorFormsValue) => {
             this.wizardValue = wizardValue;
         }, () => {});
+    }
+
+    _editStep(event: MouseEvent, stepId: string, callback: (stepId: string) => void): void {
+        event.preventDefault();
+        callback(stepId);
     }
 }