+++ libs/core/src/lib/fundamental-ngx.module.ts
@@ -73,6 +73,7 @@ import { IllustratedMessageModule } from '@fundamental-ngx/core/illustrated-mess
 import { MenuModule, MenuMobileModule } from '@fundamental-ngx/core/menu';
 import { TabsModule } from '@fundamental-ngx/core/tabs';
 import { ResizableCardLayoutModule } from '@fundamental-ngx/core/resizable-card-layout';
+import { UploadCollectionModule } from '@fundamental-ngx/core/upload-collection';
 
 import { PaginationModule } from '@fundamental-ngx/core/pagination';
 import { SelectMobileModule, SelectModule } from '@fundamental-ngx/core/select';
@@ -181,6 +182,7 @@ import { TimelineModule } from '@fundamental-ngx/core/timeline';
         ResizableCardLayoutModule,
         MessagePageModule,
         VerticalNavigationModule,
+        UploadCollectionModule,
         MicroProcessFlowModule,
         TimelineModule
     ],